CheckBox Util move and minor updates
