+++ test/controllers/review.controller.unit.ts
@@ -50,7 +50,7 @@ const SERVICE_UNAVAILABLE_TEXT = "Sorry, there is a problem with the service";
 const PAYMENT_URL = "/payment/1234";
 const PAGE_HEADING = "Submit the confirmation statement";
 const ERROR_PAGE_HEADING = "Service offline - File a confirmation statement";
-const COSTS_TEXT = "You will need to pay a fee of £34";
+const COSTS_TEXT = "You will need to pay a fee of &pound;34";
 const CONFIRMATION_STATEMENT_TEXT = "By continuing, you confirm that all information required to be delivered by the company pursuant to";
 const CONFIRMATION_STATEMENT_ECCT_TEXT = "confirm that all information required to be delivered by the company pursuant to";
 // const LAWFUL_ACTIVITY_STATEMENT_TEXT = "confirm that the intended future activities of the company are lawful";
